refactor: rename and move components
