+++ packages/twenty-front/src/pages/settings/data-model/SettingsObjectFieldEdit.tsx
@@ -11,11 +11,11 @@ import { isLabelIdentifierField } from '@/object-metadata/utils/isLabelIdentifie
 import { SaveAndCancelButtons } from '@/settings/components/SaveAndCancelButtons/SaveAndCancelButtons';
 import { SettingsHeaderContainer } from '@/settings/components/SettingsHeaderContainer';
 import { SettingsPageContainer } from '@/settings/components/SettingsPageContainer';
-import { SettingsDataModelFieldPreviewFormCard } from '@/settings/data-model/components/SettingsDataModelFieldPreviewFormCard';
-import { SettingsDataModelFieldTypeSelect } from '@/settings/data-model/components/SettingsDataModelFieldTypeSelect';
 import { SettingsObjectFieldCurrencyFormValues } from '@/settings/data-model/components/SettingsObjectFieldCurrencyForm';
 import { SettingsObjectFieldFormSection } from '@/settings/data-model/components/SettingsObjectFieldFormSection';
-import { useFieldMetadataForm } from '@/settings/data-model/hooks/useFieldMetadataForm';
+import { SettingsDataModelFieldSettingsFormCard } from '@/settings/data-model/fields/forms/components/SettingsDataModelFieldSettingsFormCard';
+import { SettingsDataModelFieldTypeSelect } from '@/settings/data-model/fields/forms/components/SettingsDataModelFieldTypeSelect';
+import { useFieldMetadataForm } from '@/settings/data-model/fields/forms/hooks/useFieldMetadataForm';
 import { AppPath } from '@/types/AppPath';
 import { IconArchive, IconSettings } from '@/ui/display/icon';
 import { H2Title } from '@/ui/display/typography/components/H2Title';
@@ -232,7 +232,7 @@ export const SettingsObjectFieldEdit = () => {
             onChange={handleFormChange}
             value={formValues.type}
           />
-          <SettingsDataModelFieldPreviewFormCard
+          <SettingsDataModelFieldSettingsFormCard
             disableCurrencyForm
             fieldMetadataItem={{
               icon: formValues.icon,
